backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
bcrypt
pydantic
python-dotenv
